separate out renderTopBind to Components.GHCCore
